Break response tests up
